requirements.txt
requests~=2.31.0
Flask~=2.3.2
python-dotenv~=1.0.0
pip~=22.3.1
argparse~=1.4.0
wheel~=0.38.4
Jinja2~=3.1.2
MarkupSafe~=2.1.3
Werkzeug~=2.3.6
click~=8.1.6
blinker~=1.6.2
itsdangerous~=2.1.2
certifi~=2023.7.22
idna~=3.4
urllib3~=2.0.4
setuptools~=65.5.1
future~=0.18.3